xml/parsing/manakai/pis-2.dat

#data
<!DOCTYPE a[
<?x?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<?x y?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x y?>
| <{} a>

#data
<!DOCTYPE a[
<?x?>
<?y?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
3;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
|   <?y ?>
| <{} a>

#data
<!DOCTYPE a[
<?x?>
]>
<?y?>
<a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <?y ?>
| <{} a>

#data
<!DOCTYPE a[
<?x ?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<?x ??y?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ??y?>
| <{} a>

#data
<!DOCTYPE a[
<?x
#errors
1;1;no XML decl;;;s
2;4;parser:EOF;;;m
2;4;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<?x\u0020
#errors
1;1;no XML decl;;;s
2;5;parser:EOF;;;m
2;5;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?x y
#errors
1;1;no XML decl;;;s
2;6;parser:EOF;;;m
2;6;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?x ?
#errors
1;1;no XML decl;;;s
2;6;parser:EOF;;;m
2;6;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?
#errors
1;1;no XML decl;;;s
2;3;parser:EOF;;;m
2;3;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?>
#errors
1;1;no XML decl;;;s
2;4;parser:EOF;;;m
2;4;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?>
]><a/>
#errors
1;1;no XML decl;;;s
3;7;parser:EOF;;;m
3;7;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<??>
<?x?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;bare pio;;;m
3;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<?x?>
%b;
<?y?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
3;1;entity not declared;;%b;;m
3;1;stop processing;;;i
4;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
|   <?y ?>
| <{} a>

#data
<?xml version="1.0" standalone="yes"?>
<!DOCTYPE a[
<?x?>
%b;
<?y?>
]><a/>
#errors
4;1;entity not declared;;%b;;m
3;1;xml:dtd:pi;;;w
5;1;xml:dtd:pi;;;w
#xml-standalone
true
#document
| <!DOCTYPE a>
|   <?x ?>
|   <?y ?>
| <{} a>

#data
<!DOCTYPE a[
  <?a+b ?>
]><a/>
#errors
1;1;no XML decl;;;s
2;3;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?a+b ?>
| <{} a>

#data
<!DOCTYPE a[
  <?a+b ?>
]><a/>
#checker ForValidator
#errors
1;1;no XML decl;;;s
2;3;xml:not name;;a+b;m
2;3;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?a+b ?>
| <{} a>

#data
<!DOCTYPE a[
  <?xml ?>
]><a/>
#errors
1;1;no XML decl;;;s
2;3;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?xml ?>
| <{} a>

#data
<!DOCTYPE a[
  <?xml ?>
]><a/>
#checker ForValidator
#errors
1;1;no XML decl;;;s
2;3;xml:pi:target:xml;;xml;m
2;3;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?xml ?>
| <{} a>

#data
<!DOCTYPE a[
  <?xMl ?>
]><a/>
#checker ForValidator
#errors
1;1;no XML decl;;;s
2;3;xml:pi:target:xml;;xMl;m
2;3;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?xMl ?>
| <{} a>

#data
<!DOCTYPE a[
  <?XML ?>
]><a/>
#checker ForValidator
#errors
1;1;no XML decl;;;s
2;3;xml:pi:target:xml;;XML;m
2;3;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?XML ?>
| <{} a>

#data
<?ho+ge?><a/>
#errors
1;1;no XML decl;;;s
#document
| <?ho+ge ?>
| <{} a>

#data
<!DOCTYPE a><?ho+ge?><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <?ho+ge ?>
| <{} a>

#data
<a/><?ho+ge?>
#errors
1;1;no XML decl;;;s
#document
| <{} a>
| <?ho+ge ?>

#data
<a><?ho+ge?></a>
#errors
1;1;no XML decl;;;s
#document
| <{} a>
|   <?ho+ge ?>

#data
<?XML?><a/>
#errors
1;1;no XML decl;;;s
#document
| <?XML ?>
| <{} a>

#data
<!DOCTYPE a><?xml?><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <?xml ?>
| <{} a>

#data
<a/><?xml?>
#errors
1;1;no XML decl;;;s
#document
| <{} a>
| <?xml ?>

#data
<a><?xml?></a>
#errors
1;1;no XML decl;;;s
#document
| <{} a>
|   <?xml ?>

#data
<?ho+ge?><a/>
#checker ForValidator
#errors
1;1;no XML decl;;;s
1;1;xml:not name;;ho+ge;m
#document
| <?ho+ge ?>
| <{} a>

#data
<!DOCTYPE a><?ho+ge?><a/>
#checker ForValidator
#errors
1;1;no XML decl;;;s
1;13;xml:not name;;ho+ge;m
#document
| <!DOCTYPE a>
| <?ho+ge ?>
| <{} a>

#data
<a/><?ho+ge?>
#checker ForValidator
#errors
1;1;no XML decl;;;s
1;5;xml:not name;;ho+ge;m
#document
| <{} a>
| <?ho+ge ?>

#data
<a><?ho+ge?></a>
#checker ForValidator
#errors
1;1;no XML decl;;;s
1;4;xml:not name;;ho+ge;m
#document
| <{} a>
|   <?ho+ge ?>

#data
<?XML?><a/>
#checker ForValidator
#errors
1;1;no XML decl;;;s
1;1;xml:pi:target:xml;;XML;m
#document
| <?XML ?>
| <{} a>

#data
<!DOCTYPE a><?xml?><a/>
#checker ForValidator
#errors
1;1;no XML decl;;;s
1;13;xml:pi:target:xml;;xml;m
#document
| <!DOCTYPE a>
| <?xml ?>
| <{} a>

#data
<a/><?xml?>
#checker ForValidator
#errors
1;1;no XML decl;;;s
1;5;xml:pi:target:xml;;xml;m
#document
| <{} a>
| <?xml ?>

#data
<a><?xml?></a>
#checker ForValidator
#errors
1;1;no XML decl;;;s
1;4;xml:pi:target:xml;;xml;m
#document
| <{} a>
|   <?xml ?>
